Testing mixed read/writes with ReadWithDelaysStorage and S3Storage
